requirements.txt
aiohttp==3.11.11
attrs==25.3.0
catalystwan==0.40.2
chromadb==0.4.24
cohere==5.15.0
dnacentersdk==2.9.1
elasticsearch==8.8.0
httpx==0.28.1
keybert==0.9.0
langchain==0.3.25
meraki==1.53.0
openai==0.27.0
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-instrumentation-fastapi==0.55b0
opentelemetry-instrumentation-requests==0.55b0
opentelemetry-instrumentation-httpx==0.55b0
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
pydantic==2.11.5
python-dotenv==1.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
sentence-transformers==4.1.0
structlog==25.3.0
sumy==0.11.0
transformers==4.51.3
yake==0.4.8